Move PlayOneShot logic to another system
